# Pay Julia JIT latency price once per day but so hacky