--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,4 @@
 .vscode
 **/*.o
+**/Manifest.toml
 bazel-*