--- a/Project.toml
+++ b/Project.toml
@@ -1,0 +1,4 @@
+name = "AoC2023"
+uuid = "6a329b9a-879f-4671-b49f-bf1b928a466a"
+authors = ["all authors"]
+version = "0.1.0"