--- a/src/julia/00_test.jl
+++ b/src/julia/00_test.jl
@@ -1,2 +1,6 @@
-println("hello world")
-println("solution 2")
+function main(path)
+    println("hello world")
+    println("solution 2")
+end
+
+(abspath(PROGRAM_FILE) == @__FILE__) && main(ARGS[1])