--- a/src/julia/02_moelf.jl
+++ b/src/julia/02_moelf.jl
@@ -1,6 +1,3 @@
-const INPUT_PATH = ARGS[1]
-const ALL_LINES = readlines(INPUT_PATH)
-
 function color_dict(a::RegexMatch)
     num, color = a.captures
     Dict(color => parse(Int, num))
@@ -18,10 +15,15 @@ function part1_pred(line)
     return dict["red"] <= 12 && dict["green"] <= 13 && dict["blue"] <= 14
 end
 
-p1 = sum(findall(part1_pred, ALL_LINES))
-println(p1)
 
 # Part II
 part2(line) = line |> line_max |> values |> prod
-p2 = sum(part2, ALL_LINES)
-println(p2)
+
+function main(path)
+    lines = readlines(path)
+    p1 = sum(findall(part1_pred, lines))
+    println(p1)
+    println(sum(part2, lines))
+end
+
+(abspath(PROGRAM_FILE) == @__FILE__) && main(ARGS[1])