## Main
```
> time julia --startup-file=no --project=. runtests.jl julia

Test Summary:                                 | Pass  Total  Time
Day 04                                        |    8      8  2.8s
  julia   04_moelf.jl          04_example.txt |    2      2  0.7s
  julia   04_moelf.jl          04_moelf.txt   |    2      2  0.7s
  julia   04_moelf.jl          04_rk.txt      |    2      2  0.7s
  julia   04_moelf.jl          04_ylks.txt    |    2      2  0.7s
________________________________________________________
Executed in   14.67 secs    fish           external
   usr time   15.00 secs    0.00 micros   15.00 secs
   sys time    6.27 secs  986.00 micros    6.27 secs
```

## This PR
```
> time julia --startup-file=no runtests.jl julia
  Activating project at `~/Documents/github/AoC2023`

Test Summary:                                 | Pass  Total  Time
Day 04                                        |    8      8  0.2s
  julia   04_moelf.jl          04_example.txt |    2      2  0.2s
  julia   04_moelf.jl          04_moelf.txt   |    2      2  0.0s
  julia   04_moelf.jl          04_rk.txt      |    2      2  0.0s
  julia   04_moelf.jl          04_ylks.txt    |    2      2  0.0s

________________________________________________________
Executed in    2.56 secs    fish           external
   usr time    2.52 secs  763.00 micros    2.52 secs
   sys time    0.39 secs  136.00 micros    0.39 secs
```