@RongkunWang 2023-12-04
> Don't over-do AOC lol..
You could use stages of the ci, compile in first stage and pass the artifacts for the second stage for tests.

If the ci time is too long you can spawn different days into different jobs, or one job per input. But I'm not sure if ci could read from directory dynamically and spawn jobs, so maybe you'd have to manually change workflow every day? You could also spawn 25 jobs and in each job detects if inputs are there, if not, pass as succeeful.