@Moelf 2023-12-04
I think the cleanest way is to ask Julia submitters (i.e. me) to write every day in:

```julia
module DayN
...
end
```

but this PR seems to work albeit hacky, and it allows everyday's script to stay in individually-runnable script