@Moelf 2023-12-04
But for JIT based language, running is compiling..